fix so kafka producer metrics are not removed
